+++ roqoqo-qasm/src/interface.rs
@@ -37,7 +37,7 @@ const ALLOWED_OPERATIONS: &[&str; 11] = &[
 ];
 
 // Operations that are ignored when looking for a QASM definition
-const NO_DEFINITION_REQUIRED_OPERATIONS: &[&str; 10] = &[
+const NO_DEFINITION_REQUIRED_OPERATIONS: &[&str; 11] = &[
     "SingleQubitGate",
     "DefinitionFloat",
     "DefinitionUsize",
@@ -48,6 +48,7 @@ const NO_DEFINITION_REQUIRED_OPERATIONS: &[&str; 10] = &[
     "PragmaGlobalPhase",
     "PragmaRepeatedMeasurement",
     "MeasureQubit",
+    "PragmaLoop",
 ];
 
 /// Calls the parsing function of the VariableGatherer, if present.
@@ -770,7 +771,8 @@ pub fn call_operation(
             }
         },
         Operation::PragmaGlobalPhase(op) => match qasm_version {
-            QasmVersion::V3point0(_) => Ok(format!("gphase {};", op.phase(),)),
+            QasmVersion::V3point0(Qasm3Dialect::Roqoqo) => Ok(format!("gphase {};", op.phase(),)),
+            QasmVersion::V3point0(Qasm3Dialect::Vanilla) => Ok(format!("gphase {};", op.phase(),)),
             _ => {
                 if ALLOWED_OPERATIONS.contains(&operation.hqslang()) {
                     Ok("".to_string())
@@ -783,16 +785,13 @@ pub fn call_operation(
             }
         },
         Operation::PragmaLoop(op) => match qasm_version {
-            QasmVersion::V2point0 => Err(RoqoqoBackendError::GenericError {
-                msg: "PragmaLoop not allowed with qasm_version 2.0".to_string(),
-            }),
             QasmVersion::V3point0(Qasm3Dialect::Roqoqo) => Ok(format!(
                 "pragma roqoqo {} {} {};",
                 op.hqslang(),
                 op.repetitions(),
                 op.circuit()
             )),
-            QasmVersion::V3point0(_) => {
+            QasmVersion::V3point0(Qasm3Dialect::Vanilla) => {
                 let mut data = "".to_string();
                 match op.repetitions() {
                     CalculatorFloat::Float(x) => {
@@ -805,6 +804,25 @@ pub fn call_operation(
                             data.push_str(format!("    {string}").as_str());
                         }
                         data.push_str("\n}");
+                        Ok(data)
+                    },
+                    CalculatorFloat::Str(x) => Err(RoqoqoBackendError::GenericError { msg: format!("Used PragmaLoop with a string {x} for repetitions and a qasm-version that is incompatible: {qasm_version:?}") })
+                }
+            }
+            _ => {
+                let mut data = "".to_string();
+                match op.repetitions() {
+                    CalculatorFloat::Float(x) => {
+                        for _ in 0_usize..(*x as usize) {
+                            let circuit_vec = match call_circuit(op.circuit(), qubit_register_name, qasm_version) {
+                                Ok(vec_str) => vec_str,
+                                Err(x) => return Err(x)
+                            };
+                            for string in circuit_vec {
+                                data.push_str(string.as_str());
+                                data.push('\n');
+                            }
+                        }
                         Ok(data)
                     },
                     CalculatorFloat::Str(x) => Err(RoqoqoBackendError::GenericError { msg: format!("Used PragmaLoop with a string {x} for repetitions and a qasm-version that is incompatible: {qasm_version:?}") })