+++ CHANGELOG.md
@@ -6,6 +6,14 @@ This changelog track changes to the qoqo qasm project starting at version 0.5.0
 
 * Added support for parametric gates using OpenQASM 3.0
 
+### 0.8.3
+
+* Bugfix for PragmaLoop in 3.0Braket and 2.0
+
+### 0.8.2
+
+* Bugfix for PragmaGlobalPhase in 3.0Braket
+
 ### 0.8.1
 
 * Updated to qoqo 1.5.1